Separate generated resources from resource_includes
