+++ .rubocop.yml
@@ -4,14 +4,7 @@ AllCops:
   Exclude:
     - 'bin/**/*'
     - 'examples/**/*'
-    - 'lib/asana/resources/attachment.rb'
-    - 'lib/asana/resources/project.rb'
-    - 'lib/asana/resources/story.rb'
-    - 'lib/asana/resources/tag.rb'
-    - 'lib/asana/resources/task.rb'
-    - 'lib/asana/resources/team.rb'
-    - 'lib/asana/resources/user.rb'
-    - 'lib/asana/resources/workspace.rb'
+    - 'lib/asana/resources/*'
     - 'spec/templates/unicorn.rb'
     - 'spec/templates/world.rb'
     - 'test.rb'
