+++ lib/asana/resource_includes/attachment_uploading.rb
