+++ lib/asana/resources.rb
@@ -1,5 +1,8 @@
-require_relative 'resources/resource'
-require_relative 'resources/collection'
+require_relative 'resource_includes/resource'
+require_relative 'resource_includes/collection'
+
+Dir[File.join(File.dirname(__FILE__), 'resource_includes', '*.rb')]
+  .each { |resource| require resource }
 
 Dir[File.join(File.dirname(__FILE__), 'resources', '*.rb')]
   .each { |resource| require resource }
